requirements.txt
fastapi
uvicorn
email
python-dotenv